requirements.txt
Babel==2.0
boto==2.38.0
cuisine==0.7.10
debtcollector==0.7.0
decorator==4.0.2
ecdsa==0.13
envassert==0.1.8.
Fabric==1.10.2
funcsigs==0.4
ipython==4.0.0
ipython-genutils==0.1.0
iso8601==0.1.10
keyring==5.4
mock==1.3.0
monotonic==0.3
msgpack-python==0.4.6
netaddr==0.7.15
netifaces==0.10.4
os-diskconfig-python-novaclient-ext==0.1.2
os-networksv2-python-novaclient-ext==0.25
os-virtual-interfacesv2-python-novaclient-ext==0.19
oslo.config==2.2.0
oslo.i18n==2.3.0
oslo.serialization==1.8.0
oslo.utils==2.2.0
paramiko==1.15.2
path.py==7.6
pbr==1.5.0
pexpect==3.3
pickleshare==0.5
prettytable==0.7.2
pycrypto==2.6.1
pyrax==1.9.4
python-keystoneclient==1.6.0
python-novaclient==2.26.0
pytz==2015.4
PyYAML==3.11
rackspace-auth-openstack==1.3
rackspace-novaclient==1.4
rax-default-network-flags-python-novaclient-ext==0.3.1
rax-scheduled-images-python-novaclient-ext==0.3.1
requests==2.7.0
simplegeneric==0.8.1
simplejson==3.8.0
six==1.9.0
stevedore==1.7.0
traitlets==4.0.0
wheel==0.24.0
wrapt==1.10.5
-e git+https://github.com/ClusterHQ/bookshelf.git#egg=bookshelf
jinja2
